requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.8.1
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92

pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
torch==2.11.0+cu128
torchaudio==2.11.0+cu128
torchvision==0.26.0+cu128
triton==3.6.0
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
urllib3==2.6.3
wheel==0.46.3
